TASK #4355 LTE网格数据管理功能增强

+++ src/main/java/com/hgicreate/rno/web/rest/LteDtDataResource.java
@@ -98,17 +98,17 @@ public class LteDtDataResource {
             Integer originFileId = 1;
             Integer flag = originFileAttrRepository.getOriginFileAttrNum();
             if(flag == null){
-                originFileAttr1.setOriginFileId(originFileId);
+                originFileAttr1.setOriginFileId((long)originFileId);
             }else {
                 originFileId = flag + 1;
-                originFileAttr1.setOriginFileId(originFileId);
+                originFileAttr1.setOriginFileId((long)originFileId);
             }
 
             //更新文件记录RNO_ORIGIN_FILE_ATTR
             originFileAttr1.setName("area_type");
             originFileAttr1.setValue(vm.getArea_type());
             originFileAttrRepository.save(originFileAttr1);
-            originFileAttr2.setOriginFileId(originFileId);
+            originFileAttr2.setOriginFileId((long)originFileId);
             originFileAttr2.setName("business_type");
             originFileAttr2.setValue(vm.getBusiness_type());
             originFileAttrRepository.save(originFileAttr2);
